cloudflaer r2 setup done

+++ src/shared/unlinkFile.ts
@@ -1,10 +1,17 @@
 import fs from 'fs'
 import path from 'path'
+import { deleteFromR2 } from '../helpers/r2.helper'
 
 const unlinkFile = (file: string) => {
-  const filePath = path.join('uploads', file)
-  if (fs.existsSync(filePath)) {
-    fs.unlinkSync(filePath)
+  if (!file) return;
+
+  if (file.startsWith('http://') || file.startsWith('https://')) {
+    deleteFromR2(file);
+  } else {
+    const filePath = path.join('uploads', file)
+    if (fs.existsSync(filePath)) {
+      fs.unlinkSync(filePath)
+    }
   }
 }
